requirements.txt
discord.py[voice]
ffmpeg
pytube
requests
openai
google-api-python-client
python-dotenv